python: Split new python binds module in to separated source files
Less thousand lines files
