+++ python/notcurses/__init__.py
@@ -14,7 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .misc import ncstrwidth, notcurses_version, notcurses_version_components
+from .notcurses import (ncstrwidth, notcurses_version,
+                        notcurses_version_components)
 
 __all__ = (
     'ncstrwidth', 'notcurses_version', 'notcurses_version_components'
